+++ database/migrations/2019_03_08_024751_create_payments_table.php
@@ -18,7 +18,7 @@ class CreatePaymentsTable extends Migration
             $table->unsignedInteger('register_id');
             $table->tinyInteger('status');
             $table->unsignedInteger('payment_method');
-            $table->double('amount_paid');
+            $table->float('amount_paid');
             $table->timestamps();
         });
     }
